// src/types/partials/MeasureReportGroup.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TMeasureReportPopulation } from '../partials/MeasureReportPopulation';
import { TQuantity } from '../partials/Quantity';
import { TMeasureReportStratifier } from '../partials/MeasureReportStratifier';

export type TMeasureReportGroup = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    code?: TCodeableConcept;
    population?: TMeasureReportPopulation[];
    measureScore?: TQuantity;
    stratifier?: TMeasureReportStratifier[];
};